+++ Sources/Mockery/AnyInvokation.swift
@@ -1,0 +1,7 @@
+/**
+ This protocol represents any kind of invocation and is used
+ to internally type erase the generic `Invokation`.
+ 
+ Original implementation: https://github.com/devxoul/Stubber
+ */
+public protocol AnyInvokation {}