+++ Sources/Mockery/MemoryAddress.swift
@@ -1,3 +1,7 @@
 import Foundation
 
+/**
+ This typealias represents a memory address. This is used by
+ `Mock` to register function invokations.
+ */
 typealias MemoryAddress = Int